Refactoring things
