+++ controllers/users/ListUsers.js
@@ -1,0 +1,6 @@
+const models = require('../../models');
+const Sequelize = require("sequelize");
+
+module.exports = function (req, res, next) {
+
+};
