+++ openapi/paths/tags.yaml
@@ -59,34 +59,36 @@ paths:
       x-operation: getTags
       parameters:
         - in: query
-          style: form
-          name: settings
-          explode: true
+          name: tags_ids
           schema:
-            type: object
-            properties:
-              tags_ids:
-                type: array
-                items:
-                  type: integer
-                  minimum: 0
-                minItems: 1
-                description: "Only consider the given tag ID(S)"
-              categories_ids:
-                type: array
-                items:
-                  type: integer
-                  minimum: 0
-                minItems: 1
-                description: "Only consider the given tag categories ID(S)"
-              state:
-                type: string
-                enum: ["default", "validated", "pending"]
-                default: "default"
-                description: "Filter the tags by their validity. The default value ('default') stands for no filtering."
-              title:
-                type: string
-                description: "Filter the tags by their text (case-insensitive comparison)"
+            type: array
+            items:
+              type: integer
+              minimum: 0
+            minItems: 1
+            description: "Only consider the given tag ID(S)"
+        - in: query
+          name: categories_ids
+          schema:
+            type: array
+            items:
+              type: integer
+              minimum: 0
+            minItems: 1
+            description: "Only consider the given tag categories ID(S)"
+        - in: query
+          name: state
+          schema:
+            type: string
+            enum: ["default", "validated", "pending"]
+            default: "default"
+            description: "Filter the tags by their validity. The default value ('default') stands for no filtering."
+        - in: query
+          name: title
+          schema:
+            type: string
+            description: "Filter the tags by their text (case-insensitive comparison)"
+            maxLength: 100
       responses:
         '200':
           description: "An array of tag"
